+++ src/apps/hedgefarm/avalanche/hedgefarm.types.ts
@@ -1,3 +1,4 @@
 export type Performance = {
-    poolName: string, averageApy: number;
+  poolName: string;
+  averageApy: number;
 };