+++ src/apps/wolf-game/ethereum/wolf-game.wool-pouch.contract-position-fetcher.ts
@@ -8,7 +8,6 @@ import { PositionTemplate } from '~app-toolkit/decorators/position-template.deco
 import { DefaultDataProps } from '~position/display.interface';
 import { ContractPositionBalance } from '~position/position-balance.interface';
 import { MetaType } from '~position/position.interface';
-import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
   DefaultContractPositionDefinition,
   GetTokenDefinitionsParams,