+++ src/apps/yield-protocol/common/yield-protocol.borrow.contract-position-fetcher.ts
@@ -11,7 +11,6 @@ import { ContractType } from '~position/contract.interface';
 import { ContractPositionBalance } from '~position/position-balance.interface';
 import { MetaType } from '~position/position.interface';
 import { GetDefinitionsParams } from '~position/template/app-token.template.types';
-import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
   GetTokenDefinitionsParams,
   GetDisplayPropsParams,